requirements.txt
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Flask==2.2.2
google-api-core==2.10.0
google-auth==2.11.0
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
greenlet==1.1.3
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
MarkupSafe==2.1.1
mysql-connector-python==8.0.30
numpy==1.23.2
pandas==1.4.4
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.40
threadpoolctl==3.1.0
urllib3==1.26.12
Werkzeug==2.2.2
